requirements.txt
beautifulsoup4==4.12.2
playwright==1.51.0
pymongo==4.6.1
python-dotenv==1.0.1 